Set up event communication through a eventBus. Which is a new Vue instance in main.js
Eg: if a child component edits the rows data, emit an even to the eventBus,
then the other components listen for that event in their created(), using eventBus.$on,
and use the event data accordingly. In the Edit view there is a test data button,
the edit view creates a new array for the data, and emits the event, then App (ie the parent),
listens for that and then updates its own data var, which is shared to its other child components.

Removed the assets folder